--- src/colour.hpp
@@ -1,9 +1,0 @@
-#ifndef RTIOW1__COLOUR_HPP_
-#define RTIOW1__COLOUR_HPP_
-
-#include <iostream>
-#include "vec3.hpp"
-
-void write_colour(std::ostream &out, colour pixel_color, int samples_per_pixel);
-
-#endif
