+++ src/hitable_list.hpp
@@ -8,23 +8,6 @@
 using std::shared_ptr;
 using std::make_shared;
 
-#ifdef USE_CUDA
-class gpu_hitable_list : public gpu_hitable
-{
- public:
-	__device__ gpu_hitable_list() = default;
-	__device__ gpu_hitable_list(gpu_hitable** l, int n)
-	{
-		list = l;
-		list_size = n;
-	}
-	gpu_hitable** list;
-	int list_size;
-
-	__device__ bool hit(const gpu_ray& r, float t_min, float t_max, gpu_hit_record& rec) const;
-};
-#endif
-
 class hitable_list : public hitable {
  public:
 	hitable_list() = default;
